Enhancement #32: Refactor old Activity classes
Moved HomeActivity.kt, ContentActivity.kt to activities/old folder
