+++ Project Code/ARSchoolBook/app/src/main/AndroidManifest.xml
@@ -32,12 +32,12 @@
             android:name=".activities.AugmentedImagesActivity"
             android:exported="true" />
         <activity
-            android:name=".activities.ContentActivity"
+            android:name=".activities.old.ContentActivity"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustResize"
             tools:ignore="LockedOrientationActivity" />
         <activity
-            android:name=".activities.HomeActivity"
+            android:name=".activities.old.HomeActivity"
             android:screenOrientation="portrait"
             tools:ignore="LockedOrientationActivity" />
         <activity
