+++ Project Code/ARSchoolBook/app/src/main/java/com/ece/nsu/spring2021/cse499/arschoolbook/adpters/HomeAdapter.kt
@@ -10,9 +10,7 @@ import android.widget.TextView
 import android.widget.Toast
 import androidx.recyclerview.widget.RecyclerView
 import com.ece.nsu.spring2021.cse499.arschoolbook.R
-import com.ece.nsu.spring2021.cse499.arschoolbook.activities.ContentActivity
 import com.ece.nsu.spring2021.cse499.arschoolbook.activities.GeContentActivity
-import com.ece.nsu.spring2021.cse499.arschoolbook.activities.GeHomeActivity
 
 class HomeAdapter(private val chapterNumbers: Array<String>,private val chapterNames: Array<String>,
                   private val context: Context) : RecyclerView.Adapter<HomeAdapter.ViewHolder>() {
