+++ Project Code/ARSchoolBook/app/src/main/res/layout/activity_class_selection.xml
@@ -7,7 +7,7 @@
     android:layout_margin="@dimen/_13sdp"
     android:elevation="@dimen/_15sdp"
     android:background="@drawable/custom_layout_background"
-    tools:context=".activities.HomeActivity">
+    tools:context=".activities.old.HomeActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
