+++ Project Code/ARSchoolBook/app/src/main/res/layout/activity_content.xml
@@ -8,7 +8,7 @@
     android:layout_margin="@dimen/_13sdp"
     android:elevation="@dimen/_15sdp"
     android:background="@drawable/custom_layout_background"
-    tools:context=".activities.ContentActivity">
+    tools:context=".activities.old.ContentActivity">
 
     <ImageButton
         android:id="@+id/backBtn"
